Take overrides from git repo, not nix store.
My understanding of this is that when a path refers to a directory containing a
`default.nix`, nix uses the derivation from that path, which refers to the `nix`
store. Therefore, the overrides referred to using that path were relative to the
path in the nix store, which causes problems. We want to refer to the locations
in this repo, so use paths that refer directly to the file we want.