Include header in AuthorisationRequest golden test
It's a bit of a trip, but if we go to the compact encoding, break out the bits
we want, decode the base64, then decode the `ByteString`s to `Value`s and put
them in one object with a header and payload, we see what we want.

Next stop: pretty printing.