Wire in dummy auth endpoint.
Add the auth API and server to the application such that we can reach its
endpoints. In the interests of getting to a PoC that demonstrates the merging of
algebras, the auth API has been reduced to a single endpoint that increments an
in-memory counter using software transactional memory (STM). `auth/increment`
increments the counter and returns the updated value.